# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: TeamFlow
  version: '1.0'
servers:
  -
    url: 'http://teamflow-be.gplans.it'
    description: 'API Server'
paths:
...
  '/api/todo/{todoId}/comment/all':
    get:
      tags:
        - comment
      summary: 'Get comments from specific todo'
      operationId: getAllTodoComments
      parameters:
        -
          name: todoId
          in: path
          description: 'todo id'
          required: true
          schema:
            type: string
            example: e14b3225-b510-47f1-bed4-bb227696098e
        -
          name: limit
          in: query
          description: 'Limit of elements'
          required: false
          schema:
            type: integer
            example: 10
        -
          name: page
          in: query
          description: 'Current page'
          required: false
          schema:
            type: integer
            example: 1
        -
          name: sortBy
          in: query
          description: 'sort by element'
          required: false
          schema:
            type: string
            example: ''
        -
          name: sortValue
          in: query
          description: 'sorting type'
          required: false
          schema:
            type: string
            example: ''
        -
          name: filterBy
          in: query
          description: 'fields to filter'
          required: false
          schema:
            type: string
            example: ''
        -
          name: filterValue
          in: query
          description: 'Values to filter'
          required: false
          schema:
            type: string
            example: ''
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentsResponse'
        401:
          description: 'Not Authorized'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        403:
          description: Forbitten
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
      security:
        -
          bearerAuth: []
  '/api/comment/{commentId}':
    get:
      tags:
        - comment
      summary: 'Get data of a single comment'
      operationId: getComment
      parameters:
        -
          name: commentId
          in: path
          description: 'model id'
          required: true
          schema:
            type: string
            example: 59e7626e-49bf-4e63-b195-f4d826af8ad8
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/CommentResponse'
                  message:
                    type: string
                type: object
        401:
          description: 'Not Authorized'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        403:
          description: Forbitten
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        404:
          description: 'Not Found'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not found'
                type: object
      security:
        -
          bearerAuth: []
    put:
      tags:
        - comment
      summary: 'Update a specific comment'
      description: ''
      operationId: updateComment
      parameters:
        -
          name: commentId
          in: path
          description: 'model id'
          required: true
          schema:
            type: string
            example: c7817f8f-3539-46c9-a5ec-b85227bb4e07
      requestBody:
        description: 'JSON with comment data'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentBodyReq'
      responses:
        200:
          description: Updated
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items: []
                  message:
                    type: string
                    example: 'Update Success'
                type: object
        400:
          description: 'Bad request'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Bad request'
                type: object
        401:
          description: 'Not Authorized'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        403:
          description: Forbitten
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        404:
          description: 'Inserted model not exists'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not found'
                type: object
        500:
          description: 'Server Error'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Insert unsuccess'
                type: object
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - comment
      summary: 'Delete an existing comment'
      description: ''
      operationId: deleteComment
      parameters:
        -
          name: commentId
          in: path
          description: 'comment id'
          required: true
          schema:
            type: string
            example: c7817f8f-3539-46c9-a5ec-b85227bb4e07
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Delete success'
                type: object
        400:
          description: 'Bad request'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Bad request'
                type: object
        401:
          description: 'Not Authorized'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        403:
          description: Forbitten
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        404:
          description: 'Not found'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not found'
                type: object
        500:
          description: 'Server Error'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Delete unsuccess'
                type: object
      security:
        -
          bearerAuth: []
  '/api/todo/{todoId}/comment':
    post:
      tags:
        - comment
      summary: 'Create new comment in a specific todo'
      description: ''
      operationId: createComment
      parameters:
        -
          name: todoId
          in: path
          description: 'todo id'
          required: true
          schema:
            type: string
            example: e14b3225-b510-47f1-bed4-bb227696098e
      requestBody:
        description: 'JSON with comment data'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentBodyReq'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                properties:
                  data:
                    properties:
                      id:
                        type: string
                        example: 1a2d4d27-43d1-4450-9ed0-5c4e7b802dae
                    type: object
                  message:
                    type: string
                    example: 'Insert success'
                type: object
        400:
          description: 'Bad request'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Bad request'
                type: object
        401:
          description: 'Not Authorized'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        403:
          description: Forbitten
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Not Authorized'
                type: object
        500:
          description: 'Server Error'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
                    example: 'Insert unsuccess'
                type: object
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    CommentResponse:
      properties:
        id:
          type: string
          example: 6ca498b0-3fb6-4094-ba16-42728383c587
        content:
          type: string
          example: 'Commento numero uno'
        created_at:
          type: string
          example: '2025-01-18 16:14:44'
        updated_at:
          type: string
          example: '2025-01-18 16:14:45'
        todo_id:
          type: string
          example: 1dcc7661-1251-456a-873c-2aa2028de9fd
        account_id:
          type: string
          example: 8a587029-80a2-4ae9-82e6-4f69f7383e63
        account_username:
          type: string
          example: test
      type: object
    CommentsResponse:
      properties:
        data:
          properties:
            current_page:
              type: number
              example: 1
            data:
              type: array
              items:
                $ref: '#/components/schemas/CommentResponse'
            first_page_url:
              type: string
              example: 'http://host/request/all?page=1'
            from:
              type: number
              example: 1
            last_page:
              type: number
              example: 1
            last_page_url:
              type: string
              example: 'http://host/request/all?page=1'
            links:
              type: array
              items:
                properties:
                  url:
                    type: string
                    format: nullable
                  label:
                    type: string
                    example: pagination.previous
                  active:
                    type: boolean
                type: object
            next_page_url:
              type: string
              format: nullable
            path:
              type: string
              example: 'http://host/request/all'
            per_page:
              type: number
              example: 10
            prev_page_url:
              type: string
              format: nullable
            to:
              type: number
              example: 1
            total:
              type: number
              example: 1
          type: object
        message:
          type: string
      type: object
    CommentBodyReq:
      properties:
        content:
          type: string
          example: 'Commento di prova'
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...